refactor: move to YukiHookAPI feat: add support for latest PA
