+++ app/.gitignore
@@ -1,2 +1,4 @@
 /build
 /release
+/src/main/assets/xposed_init
+/src/main/assets/yukihookapi_init
