+++ app/build.gradle
@@ -1,8 +1,9 @@
 plugins {
     id 'com.android.application'
     id 'kotlin-android'
-    id 'org.jetbrains.kotlin.plugin.serialization' version '1.6.0'
+    id 'org.jetbrains.kotlin.plugin.serialization' version '1.7.10'
     id 'kotlin-parcelize'
+    id 'com.google.devtools.ksp' version '1.7.0-1.0.6'
 }
 
 android {
@@ -19,6 +20,7 @@ android {
         versionCode 10
         versionName "1.2.3"
         resConfigs 'zh'
+        generatedDensities = []
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
 
@@ -46,8 +48,11 @@ dependencies {
     testImplementation 'junit:junit:4.+'
     androidTestImplementation 'androidx.test.ext:junit:1.1.3'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
+
+    implementation 'com.highcapable.yukihookapi:api:1.0.92'
     compileOnly 'de.robv.android.xposed:api:82'
-    compileOnly 'de.robv.android.xposed:api:82:sources'
+    ksp 'com.highcapable.yukihookapi:ksp-xposed:1.0.92'
+
     implementation 'com.github.vipulasri:timelineview:1.1.5'
     implementation 'org.jetbrains.kotlinx:kotlinx-serialization-json:1.3.3'
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.6.2'
