--- app/src/main/assets/xposed_init
@@ -1,1 +1,0 @@
-com.moefactory.bettermiuiexpress.hook.MainHook
