+++ app/src/main/java/com/moefactory/bettermiuiexpress/ktx/YukiHookKtx.kt
@@ -1,0 +1,10 @@
+package com.moefactory.bettermiuiexpress.ktx
+
+import android.content.Context
+import android.content.Intent
+import android.view.View
+
+val ContextType get() = Context::class.java
+val ViewType get() = View::class.java
+val IntentType get() = Intent::class.java
+val JavaListClass get() = java.util.List::class.java
